arquivo properties configurado e controller de publicador iniciado

+++ designacoes/src/main/java/com/example/repository/PublicadorRepository.java
@@ -3,6 +3,6 @@ package com.example.repository;
 import com.example.domains.Publicador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PublicadorRepository extends JpaRepository<Publicador, Integer> {
+public interface PublicadorRepository extends JpaRepository<Publicador, Long> {
 
 }
