+++ designacoes/src/main/java/com/example/repository/RelatorioRepository.java
@@ -3,5 +3,5 @@ package com.example.repository;
 import com.example.domains.Relatorio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RelatorioRepository extends JpaRepository<Relatorio, Integer> {
+public interface RelatorioRepository extends JpaRepository<Relatorio, Long> {
 }
